Pharmacokinetic Dataset of Acetaminophen

The Acetaminophen dataset contains a human experiment of pharmacokinetic data for acetaminophen and it's major metabolites acetaminophen-glucuronide and acetaminophen-sulfate.

Usage

APAP

Format

A data frame containing the following columns:

Wt: averaged weight of the study population (kg).

Dose: given oral dose of acetaminophen administered (mg/kg).

Time: time after drug administration (hr).

APAP: concentration of acetaminophen in the sample (mg/L).

AG: concentration of acetaminophen-glucuronide in the sample (mg/L).

AS: concentration of acetaminophen-sulfate in the sample (mg/L).

Study: sourced reference.

References

Zurlinden, T. J., & Reisfeld, B. (2016). Physiologically based modeling of the pharmacokinetics of acetaminophen and its major metabolites in humans using a Bayesian population approach. European journal of drug metabolism and pharmacokinetics, 41(3), 267-280.