Install MCSim

Download the latest or specific version of MCSim from the official website (https://www.gnu.org/software/mcsim/) and install it to the system directory.

install_mcsim(version = "6.0.1", directory = NULL, mxstep = 500)

Arguments

version
a character of MCSim version number.

directory
a character to assign the directory to put the MCSim files.

mxstep
a numeric value to assign the maximum number of (internally defined) steps allowed during one call to the solver.

Details

This function aims to help users install MCSim more easily. However, if you can not install it through this function. You might need to follow the instruction of MCSim and install it, manually: https://www.gnu.org/software/mcsim/mcsim.html#Installation

The default mxstp is setting to 500. The user can increase mxstp to avoid possible error return. If you meet any error when conduct sensitivity analysis, you can this function to reinstall MCSim and set the higher mxstp. The default directory to install MCSim is under /home/username (Linux), /Users/username (MacOS), and C:/Users/ (windows).

To install MCSim in Windows, be sure to install Rtools or MinGW first.

References

https://www.gnu.org/software/mcsim/

Examples

# NOT RUN {
install_mcsim(version = 6.0.1, mxstep = 10000)
# }