Download and install GNU MCSim

Download the latest or specific version of GNU MCSim from the official website (https://www.gnu.org/software/mcsim/) and install it to the system directory.

Usage

mcsim_install(version = "6.2.0", install_dir = NULL, mxstep = 5000)

mcsim_version()

set_rtools40_path()

Arguments

version

a character of version number.

install_dir

a character to assign the installed directory.

mxstep

a numeric value to assign the maximum number of (internally defined) steps allowed during one call to the solver.

Details

This function aims to help users download (source: https://ftp.gnu.org/gnu/mcsim/) and install GNU MCSim more easily. However, if you can not install it through this function. The additional way is to follow the instruction and install it manually: https://www.gnu.org/software/mcsim/mcsim.html#Installation

The default mxstp is setting to 5000. The user can increase mxstp to avoid possible error return. If you meet any error when conduct sensitivity analysis, you can use this function to re-install GNU MCSim and set the higher mxstp. The default installed directory is under /home/username (Linux), /Users/username (MacOS), and C:/Users/username/Documents (windows). To install GNU MCSim in Windows, be sure to install Rtools first. The current suggested Rtools version is 4.0.

Functions

mcsim_version(): Return the version number.

set_rtools40_path(): the function to set Rtools40 path

References

Bois, F. Y., & Maszle, D. R. (1997). MCSim: a Monte Carlo simulation program. Journal of Statistical Software, 2(9): 1–60.

https://www.gnu.org/software/mcsim/