Install MCSim

Download the latest or specific version of GNU MCSim from the official website (https://www.gnu.org/software/mcsim/) and install it to the system directory.

mcsim_install(version = "6.1.0", directory = NULL, mxstep = 500)

mcsim_version()

Arguments

version
a character of MCSim version number.

directory
a character to assign the directory to put the GNU MCSim files.

mxstep
a numeric value to assign the maximum number of (internally defined) steps allowed during one call to the solver.

Details

This function aims to help users install GNU MCSim more easily. However, if you can not install it through this function. You might need to follow the instruction of GNU MCSim and install it, manually: https://www.gnu.org/software/mcsim/mcsim.html#Installation

The default mxstp is setting to 500. The user can increase mxstp to avoid possible error return. If you meet any error when conduct sensitivity analysis, you can this function to reinstall GNU MCSim and set the higher mxstp. The default directory to install MCSim is under /home/username (Linux), /Users/username (MacOS), and C:/Users/ (windows). To install MCSim in Windows, be sure to install Rtools or MinGW first.

Functions

mcsim_version: Return the version number of GNU MCSim.

References

F.Y. Bois, and D. Maszle, 1997, MCSim: A Monte Carlo Simulation Program, Journal of Statistical Software, 2(9): 1–60.

https://www.gnu.org/software/mcsim/

Examples

# NOT RUN {
mcsim_install(version = 6.1.0, mxstep = 10000)
# }