Extended Fourier Amplitude Sensitivity Test with Random Phase Shift

Applying the extended Fourier amplitude sensitivity Test algorithm to create the numeric sequences for each parameter (Saltelli et al. 1999). Each sequence is random generated based on the random phase shift approach. It is an extension based on the fast99 function in sensitivity package.

rfast99(
  params,
  n,
  M = 4,
  omega = NULL,
  q = NULL,
  q.arg = NULL,
  replicate = 5,
  conf = 0.95
)

Arguments

params
an integer for the giving number of parameters, or a vector of character strings giving their names.

n
an integer for the sampling size.

M
an integer specifying the interference parameter. The default is 4.

omega
a vector giving the set of frequencies.

q
a vector of quantile functions names corresponding to wanted parameters distributions.

q.arg
a list of quantile functions parameters.

replicate
an integer to define the number of replication. The default is 5.

conf
the confidence level for replication confidence intervals. The default is 0.95.

Value

The returned parameter value will be stored in an array with c(model evaluation, replication, parameters).

References

Saltelli, A., Tarantola, S., & Chan, K. S. (1999). A quantitative model-independent method for global sensitivity analysis of model output. Technometrics, 41, 39-56.

Cukier, R. I., Levine, H. B., & Shuler, K. E. (1978). Nonlinear sensitivity analysis of multiparameter model systems. Journal of Computational Physics, 26, 1–42.

Examples

# Generate the parameter matrix with 20 replications
q <- "qunif"
q.arg <- list(min = 0, max = 1)

set.seed(1234)
x <- rfast99(params = 3, n = 100, replicate = 20, q = q, q.arg = q.arg)
dim(x$a) # the array of c(model evaluation, replication, parameters).

#> [1] 300  20   3

if (FALSE) {
save(x, file = "input_parameters.rda")
}