Solve PK Model Through MCSim

The solve_mcsim solves for the differential equations of time-dependent quantity/concentration in different tissues/compartments through MCSim. The output result is the 4-dimension array with c(model evaluations, replications, time-points, output variables).

solve_mcsim(x, mName, infile.name, outfile.name, n = NULL,
  setpoint.name = NULL, params = NULL, vars = NULL, time = NULL,
  condition = NULL)

generate_infile(infile.name, outfile.name, params, vars, time, condition,
  rtol = 1e-06, atol = 1e-09, n = NULL, dist = NULL,
  q.arg = NULL)

Arguments

x
a list of storing information in the defined sensitivity function.

mName
a string giving the name of the model or C file (without extension).

infile.name
a character to assign the name of input file.

outfile.name
a character to assign the name of output file.

n
a numeric to define the sample number.

setpoint.name
a character to assign the name of file for parameter matrix.

params
a character to assign the testing parameters.

vars
a character or a vector to assign the selected output(s).

time
a numeric to define the given time point(s).

condition
a character to set the specific parameter value in the input file.

rtol
an argument passed to the integrator (default 1e-6).

atol
an argument passed to the integrator (default 1e-9).

dist
a vector of distribution names corresponding to <distribution-name> in MCSim.

q.arg
a list of shape parameters in the sampling distribution (dist).

Details

This function allows users to use external data file that assigned in setpoint.name as parameter matrix. If you want to use it, be sure to define n and setpoint.name.

Functions

solve_mcsim: Numerical analysis for the PK model by MCSim.

generate_infile: Generate the MCSim input file.